Add files via upload
"Word Guesser" is an engaging and interactive text-based game that challenges players to unravel hidden words. With a wide selection of words to choose from, including "banana," "apple," "orange," "grape," and "pineapple," players are presented with a diverse range of word options to guess.

The game incorporates various features to enhance the gameplay experience. Players can select their preferred difficulty level, with options for easy, medium, and hard levels. Each difficulty level determines the number of attempts available, adding an element of strategy and decision-making to the game.

Throughout the game, players are provided with valuable information to guide their guesses. The interface displays the partially hidden word, the number of attempts remaining, the guessed letters, and the incorrect guesses. This feedback allows players to make informed decisions and adapt their strategy as they progress.

To assist players in their quest, the game offers hints in the form of additional information. Players can choose to receive a hint if they are unsure about their next move. Hints reveal one of the unguessed letters in the word, giving players a valuable clue to narrow down their options and increase their chances of success.

Scoring adds an exciting competitive element to the game. Players earn points based on the length of the word they guess correctly and the number of attempts remaining. The higher the score, the greater the accomplishment. Players can track their current score and aim to beat their personal best or compete with friends to achieve the highest score.

"Word Guesser" provides endless entertainment, engaging players in a fun and challenging word-guessing adventure. Whether you are a word enthusiast looking to expand your vocabulary or simply seeking an entertaining pastime, this game offers an enjoyable experience that tests your word-solving skills and keeps you coming back for more.